+++ swarm-derive/src/lib.rs
@@ -316,12 +316,14 @@ fn build_struct(ast: &DeriveInput, data_struct: &DataStruct) -> TokenStream {
                     self.#i.on_swarm_event(#from_swarm::ListenFailure(#listen_failure {
                             local_addr,
                             send_back_addr,
+                            error
                         }));
                     },
                     None => quote! {
                     self.#enum_n.on_swarm_event(#from_swarm::ListenFailure(#listen_failure {
                             local_addr,
                             send_back_addr,
+                            error
                         }));
                     },
                 };
@@ -781,7 +783,7 @@ fn build_struct(ast: &DeriveInput, data_struct: &DataStruct) -> TokenStream {
                         #dial_failure { peer_id, connection_id, error })
                     => { #(#on_dial_failure_stmts)* }
                     #from_swarm::ListenFailure(
-                        #listen_failure { local_addr, send_back_addr })
+                        #listen_failure { local_addr, send_back_addr, error })
                     => { #(#on_listen_failure_stmts)* }
                     #from_swarm::NewListener(
                         #new_listener { listener_id })